requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# LangChain & AI
langchain==0.1.5
langchain-community==0.0.16
langchain-google-genai==0.0.6
langchain-text-splitters==0.0.1

# Google API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0

# Database
redis==5.0.1

# Document Processing
pypdf==3.17.4
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.1
email-validator==2.1.0
python-multipart==0.0.6

# Optional: For Streamlit UI
# streamlit==1.31.0
# requests==2.31.0
